Tab component conversion to TS
